backend/requirements.txt
flask==2.3.2
flask_sqlalchemy==3.1.1
pyjwt==2.8.0
bcrypt
psycopg2-binary
Werkzeug==2.3.8